fire-detection-ai/requirements.txt
tensorflow-cpu
keras
opencv-python
numpy
matplotlib
